Refactored utilities
