+++ src/main.cpp
@@ -2,7 +2,6 @@
 
 #include "config/Config.hpp"
 #include "option/Option.hpp"
-#include "time/Time.hpp"
 
 #include <clocale>
 #include <iostream>
