+++ src/option/impl/RemoveOption.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "../Option.hpp"
-#include "../../regex/RegexHelper.hpp"
-#include "../../time/Time.hpp"
+#include "../../utils/RegexUtils.hpp"
+#include "../../utils/TimeUtils.hpp"
 #include "../../entry/Entry.hpp"
 #include <iostream>
 
@@ -17,7 +17,7 @@ public:
             return;
         }
 
-        if (!Okane::matchesIndex(args.at(0)))
+        if (!Okane::Regex::matchesIndex(args.at(0)))
         {
             std::cout << "Please enter a valid index (0, 1 or 20)";
             return;
@@ -25,12 +25,12 @@ public:
 
         size_t index = std::stoul(args.at(0));
 
-        std::string month = Okane::toStringFMT(Okane::getCurrentTime(), "%m");
-        std::string year = Okane::toStringFMT(Okane::getCurrentTime(), "%Y");
+        std::string month = Okane::Time::toStringFMT(Okane::Time::getCurrentTime(), "%m");
+        std::string year = Okane::Time::toStringFMT(Okane::Time::getCurrentTime(), "%Y");
 
         if (args.size() >= 2)
         {
-            auto monthById = Okane::getMonthFromId(args.at(1));
+            auto monthById = Okane::Time::getMonthFromId(args.at(1));
 
             if (!monthById.has_value())
             {
@@ -47,7 +47,7 @@ public:
 
             auto yearArg = args.at(2);
 
-            if (!Okane::matchesYear(yearArg))
+            if (!Okane::Regex::matchesYear(yearArg))
             {
                 std::cout << "Please provide a valid Year (2022 or 2023)" << std::endl;
                 return;
