+++ src/option/impl/StatusOption.hpp
@@ -3,10 +3,9 @@
 #include "rang.hpp"
 
 #include "../Option.hpp"
-#include "../../time/Time.hpp"
+#include "../../utils/OkaneUtils.hpp"
 #include "../../entry/Entry.hpp"
 #include "../../table/TableView.hpp"
-#include "../../regex/RegexHelper.hpp"
 #include "../../config/Config.hpp"
 
 #include <iostream>
@@ -17,12 +16,12 @@ class StatusOption : public Option
 public:
     void execute(const std::vector<std::string> &args) override
     {
-        std::string month = Okane::toStringFMT(Okane::getCurrentTime(), "%m");
-        std::string year = Okane::toStringFMT(Okane::getCurrentTime(), "%Y");
+        std::string month = Okane::Time::toStringFMT(Okane::Time::getCurrentTime(), "%m");
+        std::string year = Okane::Time::toStringFMT(Okane::Time::getCurrentTime(), "%Y");
 
         if (args.size() == 1)
         {
-            auto monthById = Okane::getMonthFromId(args.at(0));
+            auto monthById = Okane::Time::getMonthFromId(args.at(0));
 
             if (!monthById.has_value())
             {
@@ -34,7 +33,7 @@ public:
         }
         else if (args.size() > 1)
         {
-            auto monthById = Okane::getMonthFromId(args.at(0));
+            auto monthById = Okane::Time::getMonthFromId(args.at(0));
 
             if (!monthById.has_value())
             {
@@ -46,7 +45,7 @@ public:
 
             auto yearArg = args.at(1);
 
-            if (!Okane::matchesYear(yearArg))
+            if (!Okane::Regex::matchesYear(yearArg))
             {
                 std::cout << "Please provide a valid Year (2022 or 2023)" << std::endl;
                 return;
@@ -66,9 +65,9 @@ public:
         TableView tableView;
 
         tableView << "Balance";
-        tableView << TableView::to_string(monthEntry->getIncome()) + " " + Config::appConfig.currency;
-        tableView << TableView::to_string(monthEntry->getExpenses()) + " " + Config::appConfig.currency;
-        tableView << TableView::to_string(monthEntry->getBalance()) + " " + Config::appConfig.currency;
+        tableView << Okane::String::toString(monthEntry->getIncome()) + " " + Config::appConfig.currency;
+        tableView << Okane::String::toString(monthEntry->getExpenses()) + " " + Config::appConfig.currency;
+        tableView << Okane::String::toString(monthEntry->getBalance()) + " " + Config::appConfig.currency;
 
         tableView.print();
     }
